+++ packages/brick_sqlite/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Unreleased
 
+* When deleting many associations from a parent, remove the association in the joins table but do not delete the instance. This is only applicable to non-final instance fields. (#112)
+
 ## 0.1.6+1
 
 * Fix: when non-SQLite providerArgs are provided in a query, false is no longer returned from `SqliteProvider#exists`
@@ -9,7 +11,6 @@
 * Internal: Change `SqliteAdpater#fieldsToSqliteColumns` type from `Map<String, Map<String, dynamic>>` to `Map<String, RuntimeSqliteColumnDefinition>`. Using such a dynamic type option will lead to inconsistency when accessing the values.
 * `SqliteAdapter#primaryKeyByUniqueColumn` will return `instance?.primaryKey` instead of null when no `@Sqlite(unique: true)` fields exist.
 * Internal: Refactor organization of files: `SqliteProvider`, `SqliteAdapter`, `SqliteModelDictionary` are separated. `lib/sqlite.dart` is now a barrel file with the same exports.
-* When deleting many associations from a parent, remove the association in the joins table but do not delete the instance (resolves #112)
 
 ## 0.1.5+1